refactor: `GalleryImagePicker` 로직 범용화
- `ProfileImagePicker`를 `ImagePicker`로 이름을 변경하고 `ui.common.component.image` 패키지로 이동
- `ImagePicker`에 다중 선택 지원을 위한 `allowMultiple`, `selectionLimit` 파라미터 추가
- `handleImagePickerKMPCroppedImage` 함수를 `ui.common.component.image` 패키지로 이동 및 `expect/actual` 구조 재정의
