+++ mobile/composeApp/src/commonMain/composeResources/values/strings.xml
@@ -186,15 +186,16 @@
     <string name="badge_achieved_date">획득한 날: %1$s</string>
     <string name="badge_progress_rate">달성률: %1$s%</string>
 
+    <string name="image_crop_failed">이미지 자르기에 실패했습니다\n다른 이미지를 선택해보세요</string>
+    <string name="image_processing_failed">이미지 처리에 실패했습니다\n다른 이미지를 선택해보세요</string>
+    <string name="image_upload_failed">이미지 업로드에 실패했습니다\n나중에 다시 시도하세요</string>
+    <string name="image_size_failed">이미지 파일 크기를 확인할 수 없습니다</string>
+    <string name="image_selection_failed">이미지 선택 중 오류가 발생했습니다</string>
+
     <string name="setting_main_title">설정</string>
     <string name="setting_main_sign_up_date">가입한 날 : %1$s</string>
     <string name="setting_icon_description">설정 아이콘</string>
     <string name="setting_edit_profile_image">프로필 사진 수정</string>
-    <string name="setting_edit_profile_image_crop_failed">이미지 자르기에 실패했습니다\n다른 이미지를 선택해보세요</string>
-    <string name="setting_edit_profile_image_processing_failed">이미지 처리에 실패했습니다\n다른 이미지를 선택해보세요</string>
-    <string name="setting_edit_profile_image_upload_failed">이미지 업로드에 실패했습니다\n나중에 다시 시도하세요</string>
-    <string name="setting_edit_profile_get_image_size_failed">이미지 파일 크기를 확인할 수 없습니다</string>
-    <string name="setting_edit_profile_image_selection_failed">이미지 선택 중 오류가 발생했습니다</string>
     <string name="setting_edit_nickname">닉네임 수정</string>
     <string name="setting_edit_nickname_dialog_hint">새 닉네임을 입력해 주세요</string>
     <string name="setting_edited_nickname_alert">닉네임이 '%1$s'(으)로 변경되었습니다.</string>
