+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/common/component/image/ImagePicker.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.ui.setting
+package com.yagubogu.ui.common.component.image
 
 import androidx.compose.runtime.Composable
 import co.touchlab.kermit.Logger
@@ -10,34 +10,36 @@ import io.github.ismoy.imagepickerkmp.domain.models.GalleryPhotoResult
 import io.github.ismoy.imagepickerkmp.domain.models.MimeType.Companion.ALL_SUPPORTED_TYPES
 import io.github.ismoy.imagepickerkmp.presentation.ui.components.GalleryPickerLauncher
 import yagubogu.composeapp.generated.resources.Res
-import yagubogu.composeapp.generated.resources.setting_edit_profile_image_selection_failed
+import yagubogu.composeapp.generated.resources.image_selection_failed
 
 @Composable
-fun ProfileImagePicker(
-    onPhotosSelected: (String) -> Unit,
+fun ImagePicker(
+    allowMultiple: Boolean = false,
+    selectionLimit: Long = 1,
+    onPhotosSelected: (List<String>) -> Unit,
     onError: (UiText) -> Unit,
     onClosePicker: () -> Unit,
 ) {
-    val logger: Logger = Logger.withTag("ProfileImagePicker")
-    logger.d { "ProfileImagePicker 열림" }
+    val logger: Logger = Logger.withTag("ImagePicker")
+    logger.d { "ImagePicker 열림" }
     GalleryPickerLauncher(
-        allowMultiple = false,
+        allowMultiple = allowMultiple,
+        selectionLimit = selectionLimit,
         mimeTypes = ALL_SUPPORTED_TYPES,
         onPhotosSelected = { photos: List<GalleryPhotoResult> ->
             logger.d { "onPhotosSelected, 사진 개수: ${photos.size}" }
             onClosePicker()
 
-            val photo: GalleryPhotoResult? = photos.firstOrNull()
-            if (photo == null) {
+            if (photos.isEmpty()) {
                 logger.w { "선택된 사진이 없습니다" }
                 return@GalleryPickerLauncher
             }
-            onPhotosSelected(photo.uri)
+            onPhotosSelected(photos.map { it.uri })
             onClosePicker()
         },
         onError = { exception: Exception ->
             logger.e(exception) { "GalleryPicker 에러 발생" }
-            onError(UiText.StringRes(Res.string.setting_edit_profile_image_selection_failed))
+            onError(UiText.StringRes(Res.string.image_selection_failed))
             onClosePicker()
         },
         onDismiss = {
