+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/setting/SettingViewModel.kt
@@ -8,6 +8,7 @@ import com.yagubogu.data.repository.member.MemberRepository
 import com.yagubogu.data.repository.member.NicknameUpdateError
 import com.yagubogu.data.repository.member.toNicknameUpdateError
 import com.yagubogu.data.repository.thirdparty.ThirdPartyRepository
+import com.yagubogu.ui.common.component.image.handleImagePickerKMPCroppedImage
 import com.yagubogu.ui.mapper.toUiModel
 import com.yagubogu.ui.setting.model.MemberInfoItem
 import com.yagubogu.ui.setting.model.PresignedUrlCompleteItem
@@ -24,6 +25,8 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.datetime.LocalDate
 import yagubogu.composeapp.generated.resources.Res
+import yagubogu.composeapp.generated.resources.image_processing_failed
+import yagubogu.composeapp.generated.resources.image_upload_failed
 import yagubogu.composeapp.generated.resources.setting_edit_nickname_duplicate
 import yagubogu.composeapp.generated.resources.setting_edit_nickname_invalid_format
 import yagubogu.composeapp.generated.resources.setting_edit_nickname_member_not_found
@@ -32,8 +35,6 @@ import yagubogu.composeapp.generated.resources.setting_edit_nickname_no_permissi
 import yagubogu.composeapp.generated.resources.setting_edit_nickname_server_error
 import yagubogu.composeapp.generated.resources.setting_edit_nickname_too_long
 import yagubogu.composeapp.generated.resources.setting_edit_nickname_unknown_error
-import yagubogu.composeapp.generated.resources.setting_edit_profile_image_processing_failed
-import yagubogu.composeapp.generated.resources.setting_edit_profile_image_upload_failed
 import kotlin.time.Clock
 
 class SettingViewModel(
@@ -108,16 +109,16 @@ class SettingViewModel(
                 handleImagePickerKMPCroppedImage(
                     sourceImageUri = uri,
                     onUploadFailure = {
-                        emitProfileError(UiText.StringRes(Res.string.setting_edit_profile_image_upload_failed))
+                        emitProfileError(UiText.StringRes(Res.string.image_upload_failed))
                     },
                     onProcessingFailure = {
-                        emitProfileError(UiText.StringRes(Res.string.setting_edit_profile_image_processing_failed))
+                        emitProfileError(UiText.StringRes(Res.string.image_processing_failed))
                     },
                     onProfileImageUpload = ::uploadProfileImage,
                 )
             }.onFailure { exception ->
                 logger.e(exception) { "이미지 처리 과정 중 예외 발생" }
-                emitProfileError(UiText.StringRes(Res.string.setting_edit_profile_image_processing_failed))
+                emitProfileError(UiText.StringRes(Res.string.image_processing_failed))
             }
         }
     }
